change scheduler
